Organize UI into subfolders; update paths
[melhoria na organização das dos arquivos ui] Move UI files into per-page directories (ui/popup, ui/options, ui/projects, ui/dashboard) and update all references. Updated manifest (default_popup, options_page), HTML links and navigation URLs, and JS import paths (adjusted relative paths to ../../ where needed). Also added minor behavior fixes in options.js (preserve panel hash via history.replaceState and ensure a default active panel), updated README and a comment in core/entities.mjs to reflect new locations.
