+++ manifest.json
@@ -15,14 +15,14 @@
   },
   "action": {
     "default_title": "ICipó",
-    "default_popup": "ui/popup.html",
+    "default_popup": "ui/popup/popup.html",
     "default_icon": {
       "16": "img/favicon-16x16.png",
       "48": "img/favicon-48x48.png",
       "128": "img/favicon-128x128.png"
     }
   },
-  "options_page": "ui/options.html",
+  "options_page": "ui/options/options.html",
   "host_permissions": ["<all_urls>"],
   "content_scripts": [
     {
