+++ ui/dashboard/dashboard.html
@@ -400,11 +400,11 @@
 
       <div id="noActiveProjectNotice" class="floatingNotice hidden" role="status" aria-live="polite">
             <div class="title">Nenhum projeto ativo</div>
-            <div>Abra um projeto para carregar o dashboard. <a href="projects.html">Ir para projetos</a>.</div>
+            <div>Abra um projeto para carregar o dashboard. <a href="../projects/projects.html">Ir para projetos</a>.</div>
       </div>
       <div id="serverOfflineNotice" class="floatingNotice danger hidden" role="status" aria-live="polite">
         <div class="title">Sem conexão com o servidor</div>
-        <div>Certifique-se de que o servidor está rodando ou verique a conexão em <a href="options.html">configurações</a>.</div>
+        <div>Certifique-se de que o servidor está rodando ou verique a conexão em <a href="../options/options.html">configurações</a>.</div>
       </div>
     </div>
