+++ ui/projects/projects.html
@@ -15,7 +15,7 @@
     <!-- #endregion -->
 
 
-    <link rel="stylesheet" href="options.css" />
+    <link rel="stylesheet" href="../options/options.css" />
     <link rel="stylesheet" href="projects.css" />
     <script defer type="module" src="projects.js"></script>
   </head>
@@ -36,7 +36,7 @@
         </div>
 
         <div class="topbarActions">
-          <a class="sideLink" href="options.html">configurações</a>
+          <a class="sideLink" href="../options/options.html">configurações</a>
         </div>
       </header>
